Chapter order, new import, general refactoring
