+++ import/wpimport_story.py
@@ -1,5 +1,6 @@
 import os, sys
 import django
+from django.db.models import Count
 
 sys.path.append(os.path.abspath(os.path.join(__file__, *[os.pardir] * 2)))
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')
@@ -9,12 +10,11 @@ from django.contrib.auth import get_user_model
 User = get_user_model()
 from story.models import Story
 from story.models import Chapter
-from category.models import Category
+from story.models import StoryChapters
 
 import mysql.connector
 
 debug = False
-
 
 DBUSER="brawna"
 DBPASS="brawna"
@@ -93,9 +93,10 @@ for (ID, post_name, post_date, post_author, post_title,
                     created_at = post_date,
                     body = post_content,
                     old_brawna_id = ID,
-                    old_brawna_parent_id = 0,
-                    story = story
+                    old_brawna_parent_id = 0
                 )
+                storychapter = StoryChapters.objects.create(story = story, chapter = chapter, order = 1)
+
         else:
             if debug:
                 print ("Found children, make wp_post a Story, fetch all the retrieved 'chapter' posts into attached Chapters")
@@ -115,6 +116,7 @@ for (ID, post_name, post_date, post_author, post_title,
                 is_published = True
             )
             parent_brawna_id = ID
+            order = 0
             for (ID, post_name, post_date, post_author, post_title, post_content) in children:
                 if not Chapter.objects.filter(old_brawna_id=ID).exists():
                     chapter = Chapter.objects.create(
@@ -123,9 +125,10 @@ for (ID, post_name, post_date, post_author, post_title,
                         created_at = post_date,
                         body = post_content,
                         old_brawna_id = ID,
-                        old_brawna_parent_id = parent_brawna_id,
-                        story = story,
+                        old_brawna_parent_id = parent_brawna_id
                     )
+                    storychapter = StoryChapters.objects.create(story = story, chapter = chapter, order = order)
+                    order += 1
 
 if debug:
     print("All posts with post_parent==0 processed.")
@@ -135,20 +138,27 @@ if debug:
 # linked back to Storys created above.
 # We may need to climb the tree through parents
 #
-query = ('SELECT ID, post_name, post_date, post_author, post_title, post_parent \
+
+debug = True
+parent_tree = {}
+#go through once to create the parent_tree
+query = ('SELECT ID, post_parent \
+         FROM wp_posts WHERE post_status = "publish" AND post_author > 1 AND \
+         post_parent !=0 ORDER by ID')
+cnx.reconnect()
+cursor.execute(query)
+for (ID, post_parent) in cursor:
+    parent_tree[ID] = post_parent
+cnx.close()
+
+#Now go through the posts to add them as chapters
+query = ('SELECT ID, post_name, post_date, post_author, post_title, post_parent, post_content \
          FROM wp_posts WHERE post_status = "publish" AND post_author > 1 AND \
          post_parent!=0 ORDER by ID')
 
-debug = False
 cnx.reconnect()
 cursor.execute(query)
-
-parent_tree = {}
-
-for (ID, post_name, post_date, post_author,post_title, post_parent) in cursor:
-
-    parent_tree[ID] = post_parent
-
+for (ID, post_name, post_date, post_author,post_title, post_parent, post_content) in cursor:
     while not Story.objects.filter(old_brawna_id=post_parent).exists():
         if post_parent in parent_tree:
             post_parent = parent_tree[post_parent]
@@ -157,15 +167,16 @@ for (ID, post_name, post_date, post_author,post_title, post_parent) in cursor:
             break
 
     if post_parent == -1:
-        print(f"Did not connect {ID} to a parent, might be a sub-chapter")
+        if debug:
+            print(f"Did not connect {ID} to a parent, no entry in parent treer")
         continue
 
-    if debug:
-        print(f"Importing {ID} {post_name}, connecting to {post_parent}")
     try:
         parent = Story.objects.get(old_brawna_id=post_parent)
     except:
-        print(f"Did not connect {ID} to a parent might be a sub-chapter")
+        if debug:
+            print(f"Did not connect {ID} to a parent, could not find parent")
+        continue
 
     if not Chapter.objects.filter(old_brawna_id=ID).exists():
         chapter = Chapter.objects.create(
@@ -174,8 +185,12 @@ for (ID, post_name, post_date, post_author,post_title, post_parent) in cursor:
             created_at = post_date,
             body = post_content,
             old_brawna_id = ID,
-            old_brawna_parent_id = parent.old_brawna_id,
-            story = parent,
-    )
+            old_brawna_parent_id = parent.old_brawna_id
+        )
+        order = StoryChapters.objects.filter(story=parent).count() + 1
+        storychapter = StoryChapters.objects.create(story = parent, chapter = chapter, order = order)    
 
 cnx.close()
+
+#set has chapters true for anything with more than one chapter
+Story.objects.annotate(chapter_count=Count('chapters')).filter(chapter_count__gt=1).update(has_chapters=True);
